sdk/python/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "trueflow"
version = "0.1.0"
description = "Official Python SDK for TrueFlow Gateway"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "TrueFlow Team", email = "hello@trueflow.dev"}
]
dependencies = [
    "httpx>=0.24.0",
    "pydantic>=2.0",
]

[project.optional-dependencies]
openai = ["openai>=1.0.0"]
anthropic = ["anthropic>=0.3.0"]
langchain = ["langchain-openai>=0.1.0"]
crewai = ["crewai>=0.30.0"]
llamaindex = ["llama-index-llms-openai-like>=0.1.0"]
frameworks = [
    "langchain-openai>=0.1.0",
    "crewai>=0.30.0",
    "llama-index-llms-openai-like>=0.1.0",
]
dev = ["pytest>=7.0", "anyio[trio]", "requests"]


[tool.setuptools.packages.find]
where = ["."]
include = ["trueflow*"]

[tool.setuptools.package-data]
trueflow = ["py.typed"]
